requirements.txt
python-dotenv
langchain-core
langchain-google-genai
langchain-community
markitdown[all]
google-genai
